redefined TRE_CHAR_MAX to avoid the overflow issue

+++ hawk/lib/tre-compile.c
@@ -1753,9 +1753,9 @@ tre_make_trans(hawk_gem_t* gem, tre_pos_and_tags_t *p1, tre_pos_and_tags_t *p2,
 					int *tags;
 
 					DPRINT(("	 %2d -> %2d on %3d", p1->position, p2->position,
-					        p1->code_min));
+					        (int)p1->code_min));
 					if (p1->code_max != p1->code_min)
-						DPRINT(("-%3d", p1->code_max));
+						DPRINT(("-%3d", (int)p1->code_max));
 					tags = trans->tags;
 					if (tags)
 					{
